requirements.txt
Flask
aspose-3d
